+++ minimdo/debug/test_poly.py
