+++ minimdo/modeling/api.py
@@ -1,12 +1,12 @@
 from collections import OrderedDict
 from itertools import chain
-from datastructures.execution import sympy_fx_inputs, Component, edges_from_components, component_hash
-from datastructures.unitutils import get_unit, ureg
+from modeling.execution import sympy_fx_inputs, Component, edges_from_components, component_hash
+from modeling.unitutils import get_unit, ureg
 from compute import Var
 from graph.workflow import NEQ, EQ, OBJ, OPT, SOLVE
-from datastructures.graphutils import VAR, COMP, SOLVER, all_variables, copy_dicts, edges_to_Ein_Eout
-from datastructures.runpipeline import nestedform_to_mdao
-from datastructures.unitutils import fx_with_units
+from graph.graphutils import VAR, COMP, SOLVER, all_variables, copy_dicts, edges_to_Ein_Eout
+from solver.runpipeline import nestedform_to_mdao
+from modeling.unitutils import fx_with_units
 import numpy as np
 
 class SolverRef():
