+++ minimdo/presolver/tearing.py
