+++ minimdo/solver/assembly.py
@@ -1,9 +1,9 @@
 import numpy as np
 import openmdao.api as om
-from datastructures.graphutils import root_solver, nested_sources, VAR, namefromid
-from datastructures.nestedgraph import root_sources
+from graph.graphutils import root_solver, nested_sources, VAR, namefromid
+from graph.nestedgraph import root_sources
 from graph.workflow import EXPL, IMPL, SOLVE, OPT, EQ, NEQ, OBJ
-from datastructures.executionblocks import addexpcomp, addimpcomp, addsolver, addoptimizer, addoptfunc
+from solver.executionblocks import addexpcomp, addimpcomp, addsolver, addoptimizer, addoptfunc
 
 architecture_mappings = {
         EXPL: addexpcomp,
